Separate the input from the flags
The goal of this commit was to move the printing functions, such as the
ones that printed a number or the hex representation of a character, to
being methods of the Charmander type, as they could then (in the future)
inspect a Colours value to determine which colours to use.

But! This results in a partial move, because the Charmander value held
*both* the input stream *and* the current state values. The code works
fine, but Rust is stricter.

So, to appease it, that struct has been split up: the Charmander value
doesn't hold the input stream anymore, but instead takes it as an
argument.